ai-agent-frontend/src/views/voice/backend/requirements.txt
flask
flask-cors
requests
whisper
piper-tts
onnxruntime
soundfile
numpy
pyttsx3
